Move is_contain_html method to extractors.py
